+++ AvaloniaLearning/ViewModel/ViewModelBase/MainWindowViewModel.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 using System.ComponentModel;
-using AvaloniaApp.NavigationStore;
+using AvaloniaApp.NavStore;
 
 namespace AvaloniaApp.ViewModel
 {
@@ -8,9 +8,9 @@ namespace AvaloniaApp.ViewModel
     {
         public ViewModelBase? CurrentViewModel => _navStore.CurrentViewModel;
 
-        private readonly NavStore _navStore;
+        private readonly NavStore.NavigationStore _navStore;
 
-        public MainWindowViewModel(NavStore navStore)
+        public MainWindowViewModel(NavStore.NavigationStore navStore)
         {
             _navStore = navStore;
             _navStore.PropertyChanged += OnViewModelChanged;
